+++ test/common/tls/BUILD
@@ -176,6 +176,44 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
+    name = "client_context_config_impl_test",
+    srcs = [
+        "client_context_config_impl_test.cc",
+    ],
+    data = [
+        "//test/common/tls/ocsp/test_data:certs",
+        "//test/common/tls/test_data:certs",
+    ],
+    rbe_pool = "6gig",
+    deps = [
+        "ssl_certs_test_lib",
+        ":ssl_test_utils",
+        "//source/common/common:base64_lib",
+        "//source/common/json:json_loader_lib",
+        "//source/common/ssl:ssl_lib",
+        "//source/common/stats:isolated_store_lib",
+        "//source/common/stats:stats_lib",
+        "//source/common/tls:context_config_lib",
+        "//source/common/tls:context_lib",
+        "//source/common/tls:server_context_config_lib",
+        "//source/common/tls:server_context_lib",
+        "//test/common/tls/test_data:cert_infos",
+        "//test/mocks/init:init_mocks",
+        "//test/mocks/local_info:local_info_mocks",
+        "//test/mocks/runtime:runtime_mocks",
+        "//test/mocks/secret:secret_mocks",
+        "//test/mocks/server:factory_context_mocks",
+        "//test/mocks/ssl:ssl_mocks",
+        "//test/test_common:environment_lib",
+        "//test/test_common:logging_lib",
+        "//test/test_common:test_runtime_lib",
+        "@envoy_api//envoy/admin/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/transport_sockets/tls/v3:pkg_cc_proto",
+        "@envoy_api//envoy/type/matcher/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_test(
     name = "io_handle_bio_test",
     srcs = ["io_handle_bio_test.cc"],
     external_deps = ["ssl"],
